[ores.codegen,ores.refdata,agile] Fix 4 codegen defects; regen book_status to correct paths
Fix generator.py resolve_output_path() to use {component_include_path}/
{component_core_path} placeholders (slash form) for project-root directory
segments, while preserving the dotted form for C++ include directory names.
Update facet_catalogue.org to use the new placeholders. Fix three template
defects: service (missing stamp calls), generator (double counter increment),
repository (static auto → static const auto max). Apply all fixes to both
.mustache files and their canonical .org archetype sources.

Delete ghost files from ores.refdata.api/ and ores.refdata.core/ (no
CMakeLists.txt); regenerate all 18 book_status C++ files to correct
ores.refdata/api/ and ores.refdata/core/ locations. Accept generated output
(improvements over hand-written: include ordering, stamp, single idx counter,
static const max, no tenant_id in repository queries). Update
book_status_handler.hpp to align with changed protocol field names.
Second codegen run confirms idempotency.
