--- projects/ores.refdata.core/src/repository/book_status_entity.cpp
@@ -1,33 +1,0 @@
-/* -*- mode: c++; tab-width: 4; indent-tabs-mode: nil; c-basic-offset: 4 -*-
- *
- * Copyright (C) 2026 Marco Craveiro <marco.craveiro@gmail.com>
- *
- * This program is free software; you can redistribute it and/or modify it under
- * the terms of the GNU General Public License as published by the Free Software
- * Foundation; either version 3 of the License, or (at your option) any later
- * version.
- *
- * This program is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU General Public License for more
- * details.
- *
- * You should have received a copy of the GNU General Public License along with
- * this program; if not, write to the Free Software Foundation, Inc., 51
- * Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
- *
- */
-#include "ores.refdata.core/repository/book_status_entity.hpp"
-
-#include <ostream>
-#include <rfl.hpp>
-#include <rfl/json.hpp>
-
-namespace ores::refdata::repository {
-
-std::ostream& operator<<(std::ostream& s, const book_status_entity& v) {
-    rfl::json::write(v, s);
-    return s;
-}
-
-}
