--- projects/ores.refdata.core/src/repository/book_status_mapper.cpp
@@ -1,89 +1,0 @@
-/* -*- mode: c++; tab-width: 4; indent-tabs-mode: nil; c-basic-offset: 4 -*-
- *
- * Copyright (C) 2026 Marco Craveiro <marco.craveiro@gmail.com>
- *
- * This program is free software; you can redistribute it and/or modify it under
- * the terms of the GNU General Public License as published by the Free Software
- * Foundation; either version 3 of the License, or (at your option) any later
- * version.
- *
- * This program is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU General Public License for more
- * details.
- *
- * You should have received a copy of the GNU General Public License along with
- * this program; if not, write to the Free Software Foundation, Inc., 51
- * Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
- *
- */
-#include "ores.refdata.core/repository/book_status_mapper.hpp"
-
-#include "ores.database/repository/mapper_helpers.hpp"
-#include "ores.refdata.api/domain/book_status_json_io.hpp" // IWYU pragma: keep.
-
-namespace ores::refdata::repository {
-
-using namespace ores::logging;
-using namespace ores::database::repository;
-
-domain::book_status
-book_status_mapper::map(const book_status_entity& v) {
-    BOOST_LOG_SEV(lg(), trace) << "Mapping db entity: " << v;
-
-    domain::book_status r;
-    r.version = v.version;
-    r.code = v.code.value();
-    r.name = v.name;
-    r.description = v.description;
-    r.display_order = v.display_order;
-    r.modified_by = v.modified_by;
-    r.performed_by = v.performed_by;
-    r.change_reason_code = v.change_reason_code;
-    r.change_commentary = v.change_commentary;
-    if (!v.valid_from)
-        throw std::logic_error("Cannot map entity with null valid_from to domain object.");
-    r.recorded_at = timestamp_to_timepoint(*v.valid_from);
-
-    BOOST_LOG_SEV(lg(), trace) << "Mapped db entity. Result: " << r;
-    return r;
-}
-
-book_status_entity
-book_status_mapper::map(const domain::book_status& v) {
-    BOOST_LOG_SEV(lg(), trace) << "Mapping domain entity: " << v;
-
-    book_status_entity r;
-    r.code = v.code;
-    r.version = v.version;
-    r.name = v.name;
-    r.description = v.description;
-    r.display_order = v.display_order;
-    r.modified_by = v.modified_by;
-    r.performed_by = v.performed_by;
-    r.change_reason_code = v.change_reason_code;
-    r.change_commentary = v.change_commentary;
-
-    BOOST_LOG_SEV(lg(), trace) << "Mapped domain entity. Result: " << r;
-    return r;
-}
-
-std::vector<domain::book_status>
-book_status_mapper::map(const std::vector<book_status_entity>& v) {
-    return map_vector<book_status_entity, domain::book_status>(
-        v,
-        [](const auto& ve) { return map(ve); },
-        lg(),
-        "db entities");
-}
-
-std::vector<book_status_entity>
-book_status_mapper::map(const std::vector<domain::book_status>& v) {
-    return map_vector<domain::book_status, book_status_entity>(
-        v,
-        [](const auto& ve) { return map(ve); },
-        lg(),
-        "domain entities");
-}
-
-}
