+++ projects/ores.refdata/api/include/ores.refdata.api/domain/book_status_table_io.hpp
