+++ projects/ores.refdata/api/include/ores.refdata.api/generators/book_status_generator.hpp
