+++ projects/ores.refdata/api/src/domain/book_status_json_io.cpp
@@ -18,10 +18,11 @@
  *
  */
 #include "ores.refdata.api/domain/book_status_json_io.hpp"
-#include "ores.utility/rfl/reflectors.hpp" // IWYU pragma: keep.
+
 #include <ostream>
 #include <rfl.hpp>
 #include <rfl/json.hpp>
+#include "ores.utility/rfl/reflectors.hpp" // IWYU pragma: keep.
 
 namespace ores::refdata::domain {
