+++ projects/ores.refdata/api/src/domain/book_status_table_io.cpp
