+++ projects/ores.refdata/core/include/ores.refdata.core/repository/book_status_repository.hpp
@@ -20,13 +20,13 @@
 #ifndef ORES_REFDATA_CORE_REPOSITORY_BOOK_STATUS_REPOSITORY_HPP
 #define ORES_REFDATA_CORE_REPOSITORY_BOOK_STATUS_REPOSITORY_HPP
 
-#include "ores.database/domain/context.hpp"
-#include "ores.logging/make_logger.hpp"
-#include "ores.refdata.api/domain/book_status.hpp"
-#include "ores.refdata.core/export.hpp"
-#include <sqlgen/postgres.hpp>
 #include <string>
 #include <vector>
+#include <sqlgen/postgres.hpp>
+#include "ores.logging/make_logger.hpp"
+#include "ores.database/domain/context.hpp"
+#include "ores.refdata.api/domain/book_status.hpp"
+#include "ores.refdata.core/export.hpp"
 
 namespace ores::refdata::repository {
 
@@ -35,7 +35,8 @@ namespace ores::refdata::repository {
  */
 class ORES_REFDATA_CORE_EXPORT book_status_repository {
 private:
-    inline static std::string_view logger_name = "ores.refdata.repository.book_status_repository";
+    inline static std::string_view logger_name =
+        "ores.refdata.repository.book_status_repository";
 
     [[nodiscard]] static auto& lg() {
         using namespace ores::logging;
@@ -48,20 +49,18 @@ public:
 
     std::string sql();
 
-    void write(context ctx, const domain::book_status& status);
-    void write(context ctx, const std::vector<domain::book_status>& statuses);
+    void write(context ctx, const domain::book_status& v);
+    void write(context ctx, const std::vector<domain::book_status>& v);
 
     std::vector<domain::book_status> read_latest(context ctx);
-    std::vector<domain::book_status> read_latest(context ctx, const std::string& code);
-
-    std::vector<domain::book_status> read_all(context ctx, const std::string& code);
+    std::vector<domain::book_status>
+    read_latest(context ctx, const std::string& code);
+    std::vector<domain::book_status>
+    read_all(context ctx, const std::string& code);
 
     void remove(context ctx, const std::string& code);
-
-    /**
-     * @brief Deletes book statuses by closing their temporal validity.
-     */
     void remove(context ctx, const std::vector<std::string>& codes);
+
 };
 
 }
