+++ projects/ores.refdata/core/src/repository/book_status_entity.cpp
@@ -18,6 +18,7 @@
  *
  */
 #include "ores.refdata.core/repository/book_status_entity.hpp"
+
 #include <ostream>
 #include <rfl.hpp>
 #include <rfl/json.hpp>
