+++ projects/ores.refdata/core/src/repository/book_status_mapper.cpp
@@ -18,6 +18,7 @@
  *
  */
 #include "ores.refdata.core/repository/book_status_mapper.hpp"
+
 #include "ores.database/repository/mapper_helpers.hpp"
 #include "ores.refdata.api/domain/book_status_json_io.hpp" // IWYU pragma: keep.
 
@@ -26,7 +27,8 @@ namespace ores::refdata::repository {
 using namespace ores::logging;
 using namespace ores::database::repository;
 
-domain::book_status book_status_mapper::map(const book_status_entity& v) {
+domain::book_status
+book_status_mapper::map(const book_status_entity& v) {
     BOOST_LOG_SEV(lg(), trace) << "Mapping db entity: " << v;
 
     domain::book_status r;
@@ -39,13 +41,16 @@ domain::book_status book_status_mapper::map(const book_status_entity& v) {
     r.performed_by = v.performed_by;
     r.change_reason_code = v.change_reason_code;
     r.change_commentary = v.change_commentary;
-    r.recorded_at = timestamp_to_timepoint(v.valid_from);
+    if (!v.valid_from)
+        throw std::logic_error("Cannot map entity with null valid_from to domain object.");
+    r.recorded_at = timestamp_to_timepoint(*v.valid_from);
 
     BOOST_LOG_SEV(lg(), trace) << "Mapped db entity. Result: " << r;
     return r;
 }
 
-book_status_entity book_status_mapper::map(const domain::book_status& v) {
+book_status_entity
+book_status_mapper::map(const domain::book_status& v) {
     BOOST_LOG_SEV(lg(), trace) << "Mapping domain entity: " << v;
 
     book_status_entity r;
@@ -63,14 +68,22 @@ book_status_entity book_status_mapper::map(const domain::book_status& v) {
     return r;
 }
 
-std::vector<domain::book_status> book_status_mapper::map(const std::vector<book_status_entity>& v) {
+std::vector<domain::book_status>
+book_status_mapper::map(const std::vector<book_status_entity>& v) {
     return map_vector<book_status_entity, domain::book_status>(
-        v, [](const auto& ve) { return map(ve); }, lg(), "db entities");
+        v,
+        [](const auto& ve) { return map(ve); },
+        lg(),
+        "db entities");
 }
 
-std::vector<book_status_entity> book_status_mapper::map(const std::vector<domain::book_status>& v) {
+std::vector<book_status_entity>
+book_status_mapper::map(const std::vector<domain::book_status>& v) {
     return map_vector<domain::book_status, book_status_entity>(
-        v, [](const auto& ve) { return map(ve); }, lg(), "domain entities");
+        v,
+        [](const auto& ve) { return map(ve); },
+        lg(),
+        "domain entities");
 }
 
 }
