+++ projects/ores.refdata/core/src/repository/book_status_repository.cpp
@@ -18,12 +18,13 @@
  *
  */
 #include "ores.refdata.core/repository/book_status_repository.hpp"
-#include "ores.database/repository/bitemporal_operations.hpp"
+
+#include <sqlgen/postgres.hpp>
 #include "ores.database/repository/helpers.hpp"
+#include "ores.database/repository/bitemporal_operations.hpp"
 #include "ores.refdata.api/domain/book_status_json_io.hpp" // IWYU pragma: keep.
 #include "ores.refdata.core/repository/book_status_entity.hpp"
 #include "ores.refdata.core/repository/book_status_mapper.hpp"
-#include <sqlgen/postgres.hpp>
 
 namespace ores::refdata::repository {
 
@@ -36,83 +37,75 @@ std::string book_status_repository::sql() {
     return generate_create_table_sql<book_status_entity>(lg());
 }
 
-void book_status_repository::write(context ctx, const domain::book_status& status) {
-    BOOST_LOG_SEV(lg(), debug) << "Writing book status to database: " << status.code;
-    execute_write_query(
-        ctx, book_status_mapper::map(status), lg(), "Writing book status to database.");
+void book_status_repository::write(context ctx, const domain::book_status& v) {
+    BOOST_LOG_SEV(lg(), debug) << "Writing book status: " << v.code;
+    execute_write_query(ctx, book_status_mapper::map(v),
+        lg(), "Writing book status to database.");
 }
 
-void book_status_repository::write(context ctx, const std::vector<domain::book_status>& statuses) {
-    BOOST_LOG_SEV(lg(), debug) << "Writing book statuses to database. Count: " << statuses.size();
-    execute_write_query(
-        ctx, book_status_mapper::map(statuses), lg(), "Writing book statuses to database.");
+void book_status_repository::write(
+    context ctx, const std::vector<domain::book_status>& v) {
+    BOOST_LOG_SEV(lg(), debug) << "Writing book statuses. Count: " << v.size();
+    execute_write_query(ctx, book_status_mapper::map(v),
+        lg(), "Writing book statuses to database.");
 }
 
-std::vector<domain::book_status> book_status_repository::read_latest(context ctx) {
-    const auto max(make_timestamp(MAX_TIMESTAMP, lg()));
-    const auto tid = ctx.tenant_id().to_string();
+std::vector<domain::book_status>
+book_status_repository::read_latest(context ctx) {
+    static const auto max(make_timestamp(MAX_TIMESTAMP, lg()));
     const auto query = sqlgen::read<std::vector<book_status_entity>> |
-                       where("tenant_id"_c == tid && "valid_to"_c == max.value()) |
-                       order_by("name"_c);
+        where("valid_to"_c == max.value()) |
+        order_by("code"_c);
 
     return execute_read_query<book_status_entity, domain::book_status>(
-        ctx,
-        query,
+        ctx, query,
         [](const auto& entities) { return book_status_mapper::map(entities); },
-        lg(),
-        "Reading latest book statuses");
+        lg(), "Reading latest book statuses");
 }
 
-std::vector<domain::book_status> book_status_repository::read_latest(context ctx,
-                                                                     const std::string& code) {
-    BOOST_LOG_SEV(lg(), debug) << "Reading latest book status. Code: " << code;
-
-    const auto max(make_timestamp(MAX_TIMESTAMP, lg()));
-    const auto tid = ctx.tenant_id().to_string();
-    const auto query =
-        sqlgen::read<std::vector<book_status_entity>> |
-        where("tenant_id"_c == tid && "code"_c == code && "valid_to"_c == max.value());
-
-    return execute_read_query<book_status_entity, domain::book_status>(
-        ctx,
-        query,
-        [](const auto& entities) { return book_status_mapper::map(entities); },
-        lg(),
-        "Reading latest book status by code.");
-}
-
-std::vector<domain::book_status> book_status_repository::read_all(context ctx,
-                                                                  const std::string& code) {
-    BOOST_LOG_SEV(lg(), debug) << "Reading all book status versions. Code: " << code;
-
-    const auto tid = ctx.tenant_id().to_string();
+std::vector<domain::book_status>
+book_status_repository::read_latest(context ctx, const std::string& code) {
+    BOOST_LOG_SEV(lg(), debug) << "Reading latest book status. code: " << code;
+    static const auto max(make_timestamp(MAX_TIMESTAMP, lg()));
     const auto query = sqlgen::read<std::vector<book_status_entity>> |
-                       where("tenant_id"_c == tid && "code"_c == code) |
-                       order_by("version"_c.desc());
+        where("code"_c == code && "valid_to"_c == max.value());
 
     return execute_read_query<book_status_entity, domain::book_status>(
-        ctx,
-        query,
+        ctx, query,
         [](const auto& entities) { return book_status_mapper::map(entities); },
-        lg(),
-        "Reading all book status versions by code.");
+        lg(), "Reading latest book status by code.");
+}
+
+std::vector<domain::book_status>
+book_status_repository::read_all(context ctx, const std::string& code) {
+    BOOST_LOG_SEV(lg(), debug) << "Reading all book status versions. code: " << code;
+    const auto query = sqlgen::read<std::vector<book_status_entity>> |
+        where("code"_c == code) |
+        order_by("version"_c.desc());
+
+    return execute_read_query<book_status_entity, domain::book_status>(
+        ctx, query,
+        [](const auto& entities) { return book_status_mapper::map(entities); },
+        lg(), "Reading all book status versions by code.");
 }
 
 void book_status_repository::remove(context ctx, const std::string& code) {
-    BOOST_LOG_SEV(lg(), debug) << "Removing book status from database: " << code;
-
-    const auto max(make_timestamp(MAX_TIMESTAMP, lg()));
-    const auto tid = ctx.tenant_id().to_string();
-    const auto query =
-        sqlgen::delete_from<book_status_entity> |
-        where("tenant_id"_c == tid && "code"_c == code && "valid_to"_c == max.value());
+    BOOST_LOG_SEV(lg(), debug) << "Removing book status: " << code;
+    static const auto max(make_timestamp(MAX_TIMESTAMP, lg()));
+    const auto query = sqlgen::delete_from<book_status_entity> |
+        where("code"_c == code && "valid_to"_c == max.value());
 
     execute_delete_query(ctx, query, lg(), "Removing book status from database.");
 }
 
-void book_status_repository::remove(context ctx, const std::vector<std::string>& codes) {
-    const auto query = sqlgen::delete_from<book_status_entity> | where("code"_c.in(codes));
-    execute_delete_query(ctx, query, lg(), "batch removing book_statuses");
+void book_status_repository::remove(
+    context ctx, const std::vector<std::string>& codes) {
+    static const auto max(make_timestamp(MAX_TIMESTAMP, lg()));
+    const auto query = sqlgen::delete_from<book_status_entity> |
+        where("code"_c.in(codes) && "valid_to"_c == max.value());
+    execute_delete_query(ctx, query, lg(), "batch removing book statuses");
 }
+
+
 
 }
