+++ projects/ores.refdata/core/src/service/book_status_service.cpp
@@ -18,8 +18,9 @@
  *
  */
 #include "ores.refdata.core/service/book_status_service.hpp"
-#include "ores.service/messaging/handler_helpers.hpp"
+
 #include <stdexcept>
+#include "ores.service/messaging/handler_helpers.hpp"
 
 using ores::service::messaging::stamp;
 
@@ -35,36 +36,35 @@ std::vector<domain::book_status> book_status_service::list_statuses() {
     return repo_.read_latest(ctx_);
 }
 
-std::optional<domain::book_status> book_status_service::find_status(const std::string& code) {
-    BOOST_LOG_SEV(lg(), debug) << "Finding book status: " << code;
+std::optional<domain::book_status>
+book_status_service::get_status(const std::string& code) {
+    BOOST_LOG_SEV(lg(), debug) << "Getting book status: " << code;
     auto results = repo_.read_latest(ctx_, code);
-    if (results.empty()) {
-        return std::nullopt;
-    }
+    if (results.empty()) return std::nullopt;
     return results.front();
 }
 
-void book_status_service::save_status(const domain::book_status& status) {
-    if (status.code.empty()) {
-        throw std::invalid_argument("Party status code cannot be empty.");
-    }
-    BOOST_LOG_SEV(lg(), debug) << "Saving book status: " << status.code;
-    auto s = status;
-    stamp(s, ctx_);
-    repo_.write(ctx_, s);
-    BOOST_LOG_SEV(lg(), info) << "Saved book status: " << status.code;
+void book_status_service::save_status(const domain::book_status& v) {
+    if (v.code.empty())
+        throw std::invalid_argument("Book Status code cannot be empty.");
+    BOOST_LOG_SEV(lg(), debug) << "Saving book status: " << v.code;
+    auto t = v;
+    stamp(t, ctx_);
+    repo_.write(ctx_, t);
+    BOOST_LOG_SEV(lg(), info) << "Saved book status: " << v.code;
 }
 
-void book_status_service::save_statuses(const std::vector<domain::book_status>& statuses) {
-    for (const auto& s : statuses) {
-        if (s.code.empty())
-            throw std::invalid_argument("Book status code cannot be empty.");
-    }
-    BOOST_LOG_SEV(lg(), debug) << "Saving " << statuses.size() << " book statuses";
-    auto stamped = statuses;
-    for (auto& s : stamped)
-        stamp(s, ctx_);
-    repo_.write(ctx_, stamped);
+void book_status_service::save_statuses(
+    const std::vector<domain::book_status>& statuses) {
+    for (const auto& e : statuses)
+        if (e.code.empty())
+            throw std::invalid_argument("Book Status code cannot be empty.");
+    BOOST_LOG_SEV(lg(), debug) << "Saving " << statuses.size()
+        << " book statuses";
+    auto ts = statuses;
+    for (auto& e : ts)
+        stamp(e, ctx_);
+    repo_.write(ctx_, ts);
 }
 
 void book_status_service::remove_status(const std::string& code) {
@@ -73,11 +73,13 @@ void book_status_service::remove_status(const std::string& code) {
     BOOST_LOG_SEV(lg(), info) << "Removed book status: " << code;
 }
 
-void book_status_service::remove_statuses(const std::vector<std::string>& codes) {
+void book_status_service::remove_statuses(
+    const std::vector<std::string>& codes) {
     repo_.remove(ctx_, codes);
 }
 
-std::vector<domain::book_status> book_status_service::get_status_history(const std::string& code) {
+std::vector<domain::book_status>
+book_status_service::get_status_history(const std::string& code) {
     BOOST_LOG_SEV(lg(), debug) << "Getting history for book status: " << code;
     return repo_.read_all(ctx_, code);
 }
